add setup tests refactor Context and Test Setup
