+++ src/index.js
@@ -3,9 +3,17 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import "./locale/i18n";
+import AuthContextWrapper from "./state/AuthContextWrapper";
+import {BrowserRouter as Router} from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<App />);
+root.render(
+    <Router>
+        <AuthContextWrapper>
+            <App />
+        </AuthContextWrapper>
+    </Router>
+);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
