+++ src/pages/LoginPage.js
@@ -4,7 +4,7 @@ import ButtonWithProgress from "../components/ButtonWithProgress";
 import React, {useState, useEffect, useContext} from "react";
 import { login } from "../api/apiCalls";
 import { useTranslation } from "react-i18next";
-import {AuthContext} from "../App";
+import {AuthContext} from "../state/AuthContextWrapper";
 
 const LoginPage = (props) => {
     const [email, setEmail] = useState();
